refactor(app): extract gallery card data to a shared namespace too
